Add mock_bluetooth fixture (#75075)
