+++ tests/components/bluetooth/conftest.py
@@ -1,25 +1,1 @@
 """Tests for the bluetooth component."""
-
-import threading
-
-import pytest
-
-from tests.common import INSTANCES
-
-
-@pytest.fixture(autouse=True)
-def verify_cleanup():
-    """Verify that the test has cleaned up resources correctly."""
-    threads_before = frozenset(threading.enumerate())
-
-    yield
-
-    if len(INSTANCES) >= 2:
-        count = len(INSTANCES)
-        for inst in INSTANCES:
-            inst.stop()
-        pytest.exit(f"Detected non stopped instances ({count}), aborting test run")
-
-    threads = frozenset(threading.enumerate()) - threads_before
-    for thread in threads:
-        assert isinstance(thread, threading._DummyThread)
